requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Session==0.5.0
psycopg2-binary==2.9.9
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
itsdangerous==2.1.2
flask-discord==0.1.64
gunicorn==21.2.0 